Cli moved to separate file.
