+++ generate_plugin.py
@@ -10,6 +10,7 @@ and deployment workflow.
 
 from utils import generator_utils
 from utils.plugin_config import PluginConfig
+from utils import cli
 
 from pathlib import Path
 import platform
@@ -18,29 +19,29 @@ import time
 import shlex
 import shutil
 import sys
-import argparse
+# import argparse
 
 
 # Define CLI args.
-def parse_cli():
-    parser = argparse.ArgumentParser(description="Plugin generator for django-simple-deploy.")
-    parser.add_argument(
-        "--target-dir",
-        type=str,
-        help="Path where the new directory will be written.",
-    )
-    args = parser.parse_args()
+# def parse_cli():
+#     parser = argparse.ArgumentParser(description="Plugin generator for django-simple-deploy.")
+#     parser.add_argument(
+#         "--target-dir",
+#         type=str,
+#         help="Path where the new directory will be written.",
+#     )
+#     args = parser.parse_args()
 
-    # If provided, make sure target_dir exists before doing anything else.
-    if args.target_dir:
-        path = Path(args.target_dir)
-        if not path.exists():
-            msg = f"The path {path.as_posix()} does not exist."
-            msg += "\n  Please create this directory and run the plugin generator again,"
-            msg += "\n  or choose another location to write to."
-            sys.exit(msg)
+#     # If provided, make sure target_dir exists before doing anything else.
+#     if args.target_dir:
+#         path = Path(args.target_dir)
+#         if not path.exists():
+#             msg = f"The path {path.as_posix()} does not exist."
+#             msg += "\n  Please create this directory and run the plugin generator again,"
+#             msg += "\n  or choose another location to write to."
+#             sys.exit(msg)
 
-    return args
+#     return args
 
 
 def generate_plugin(plugin_config, args):
@@ -215,7 +216,7 @@ def generate_plugin(plugin_config, args):
 
 
 if __name__ == "__main__":
-    args = parse_cli()
+    args = cli.parse_cli()
 
     # Define empty PlugingConfig object, get required info, and generate plugin.
     plugin_config = PluginConfig
